refactor: implement modular to update password & update profile user
