+++ resources/js/Components/Features/Auth/LoginForm.vue
@@ -22,6 +22,15 @@ defineProps({
 
 <template>
     <form @submit.prevent="submit">
+        <!-- back to home link -->
+        <div class="flex items-center justify-end mt-4">
+            <Link :href="'/'"
+                class="underline text-sm text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-100 rounded-md focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 dark:focus:ring-offset-gray-800">
+            Back to home
+            </Link>
+        </div>
+        <!-- end back to home link -->
+
         <!-- input email -->
         <div>
             <InputLabel for="email" value="Email" />
@@ -50,7 +59,7 @@ defineProps({
         <!-- end check remember me -->
 
         <div class="flex items-center justify-end mt-4">
-            <!-- forgor password link -->
+            <!-- forgot password link -->
             <Link v-if="canResetPassword" :href="route('password.request')"
                 class="underline text-sm text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-100 rounded-md focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 dark:focus:ring-offset-gray-800">
             Forgot your password?
