+++ resources/js/Components/Features/Profile/UpdatePasswordForm.vue
@@ -1,0 +1,60 @@
+<script setup>
+import { ref } from 'vue';
+
+// form components
+import InputError from '@/Components/Form/InputError.vue';
+import InputLabel from '@/Components/Form/InputLabel.vue';
+import TextInput from '@/Components/Form/TextInput.vue';
+
+defineProps({
+    form: Object,
+});
+
+const passwordInput = ref(null);
+const currentPasswordInput = ref(null);
+
+defineExpose({
+    passwordInput,
+    currentPasswordInput,
+});
+</script>
+
+<template>
+    <div class="col-span-6 sm:col-span-4">
+        <InputLabel for="current_password" value="Current Password" />
+        <TextInput
+            id="current_password"
+            ref="currentPasswordInput"
+            v-model="form.current_password"
+            type="password"
+            class="mt-1 block w-full"
+            autocomplete="current-password"
+        />
+        <InputError :message="form.errors.current_password" class="mt-2" />
+    </div>
+
+    <div class="col-span-6 sm:col-span-4">
+        <InputLabel for="password" value="New Password" />
+        <TextInput
+            id="password"
+            ref="passwordInput"
+            v-model="form.password"
+            type="password"
+            class="mt-1 block w-full"
+            autocomplete="new-password"
+        />
+        <InputError :message="form.errors.password" class="mt-2" />
+    </div>
+
+    <div class="col-span-6 sm:col-span-4">
+        <InputLabel for="password_confirmation" value="Confirm Password" />
+        <TextInput
+            id="password_confirmation"
+            v-model="form.password_confirmation"
+            type="password"
+            class="mt-1 block w-full"
+            autocomplete="new-password"
+        />
+        <InputError :message="form.errors.password_confirmation" class="mt-2" />
+    </div>
+</template>
