+++ resources/js/Pages/Profile/Partials/UpdatePasswordForm.vue
@@ -1,6 +1,12 @@
 <script setup>
-import { ref } from 'vue';
-import { useForm } from '@inertiajs/vue3';
+// composables
+import { useUpdatePasswordForm } from '@/Composables/Features/Profile/useUpdatePasswordForm.js';
+
+// feature components
+import UpdatePasswordForm from '@/Components/Features/Profile/UpdatePasswordForm.vue';
+
+// form components
+import PrimaryButton from '@/Components/Form/PrimaryButton.vue';
 
 // message components
 import ActionMessage from '@/Components/Messages/ActionMessage.vue';
@@ -8,99 +14,38 @@ import ActionMessage from '@/Components/Messages/ActionMessage.vue';
 // section components
 import FormSection from '@/Components/Sections/FormSection.vue';
 
-// form components
-import InputError from '@/Components/Form/InputError.vue';
-import InputLabel from '@/Components/Form/InputLabel.vue';
-import PrimaryButton from '@/Components/Form/PrimaryButton.vue';
-import TextInput from '@/Components/Form/TextInput.vue';
-
-const passwordInput = ref(null);
-const currentPasswordInput = ref(null);
-
-const form = useForm({
-    current_password: '',
-    password: '',
-    password_confirmation: '',
-});
-
-const updatePassword = () => {
-    form.put(route('user-password.update'), {
-        errorBag: 'updatePassword',
-        preserveScroll: true,
-        onSuccess: () => form.reset(),
-        onError: () => {
-            if (form.errors.password) {
-                form.reset('password', 'password_confirmation');
-                passwordInput.value.focus();
-            }
-
-            if (form.errors.current_password) {
-                form.reset('current_password');
-                currentPasswordInput.value.focus();
-            }
-        },
-    });
-};
+const { updatePassword, form } = useUpdatePasswordForm();
 </script>
 
 <template>
     <FormSection @submitted="updatePassword">
+        <!-- Title -->
         <template #title>
             Update Password
         </template>
+        <!-- End Title -->
 
+        <!-- Description -->
         <template #description>
             Ensure your account is using a long, random password to stay secure.
         </template>
+        <!-- End Description -->
 
+        <!-- Form -->
         <template #form>
-            <div class="col-span-6 sm:col-span-4">
-                <InputLabel for="current_password" value="Current Password" />
-                <TextInput
-                    id="current_password"
-                    ref="currentPasswordInput"
-                    v-model="form.current_password"
-                    type="password"
-                    class="mt-1 block w-full"
-                    autocomplete="current-password"
-                />
-                <InputError :message="form.errors.current_password" class="mt-2" />
-            </div>
-
-            <div class="col-span-6 sm:col-span-4">
-                <InputLabel for="password" value="New Password" />
-                <TextInput
-                    id="password"
-                    ref="passwordInput"
-                    v-model="form.password"
-                    type="password"
-                    class="mt-1 block w-full"
-                    autocomplete="new-password"
-                />
-                <InputError :message="form.errors.password" class="mt-2" />
-            </div>
-
-            <div class="col-span-6 sm:col-span-4">
-                <InputLabel for="password_confirmation" value="Confirm Password" />
-                <TextInput
-                    id="password_confirmation"
-                    v-model="form.password_confirmation"
-                    type="password"
-                    class="mt-1 block w-full"
-                    autocomplete="new-password"
-                />
-                <InputError :message="form.errors.password_confirmation" class="mt-2" />
-            </div>
+            <UpdatePasswordForm :form="form"/>
         </template>
+        <!-- End Form -->
 
+        <!-- Actions -->
         <template #actions>
             <ActionMessage :on="form.recentlySuccessful" class="me-3">
                 Saved.
             </ActionMessage>
-
             <PrimaryButton :class="{ 'opacity-25': form.processing }" :disabled="form.processing">
                 Save
             </PrimaryButton>
         </template>
+        <!-- End Actions -->
     </FormSection>
 </template>
